+++ test/integration_test/matrix/methods/exp/exp_test.dart
@@ -1,0 +1,8 @@
+import 'package:ml_linalg/dtype.dart';
+
+import 'exp_test_group_factory.dart';
+
+void main() {
+  matrixExpTestGroupFactory(DType.float32);
+  matrixExpTestGroupFactory(DType.float64);
+}